+++ packages/gatsby-theme-aio/src/components/RedoclyAPIBlock/index.js
@@ -18,7 +18,13 @@ import PropTypes from 'prop-types';
 const licenseKey = process.env.GATSBY_REDOCLY_KEY;
 
 // Redocly API Block that will render the OpenAPI yaml files with Redocly TryIt feature.
-const RedoclyAPIBlock = ({ src }) => {
+const RedoclyAPIBlock = ({
+  src, width = '500px',
+  typography = 'fontFamily: `adobe-clean, "Source Sans Pro", -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, Ubuntu, "Trebuchet MS", "Lucida Grande", sans-serif`',
+  codeBlock = "tokens: { punctuation: { color: 'white' }}",
+  disableSidebar = false,
+  disableSearch = false
+}) => {
   const [isRedoclyLoading, setIsRedoclyLoading] = useState(true);
 
   let input = {};
@@ -48,16 +54,21 @@ const RedoclyAPIBlock = ({ src }) => {
     <>
       {!isRedoclyLoading && (
         <>
-          <div id="redocly_container"/>
+          <div id="redocly_container" />
 
           <script>{
             `RedoclyReferenceDocs.init(
                '${src}',
               {licenseKey: '${licenseKey}',
+               disableSidebar: ${disableSidebar}, 
+               disableSearch: ${disableSearch},
+               hideLoading: true,
                theme: {
+                ${typography ? "typography: { " + typography + "}," : ''}
                 rightPanel: {
-                  width: '500px',
+                  width: '${width}',
                   },
+                  ${codeBlock ? "codeBlock: { " + codeBlock + "}," : ''}
                 },
               },
               document.querySelector('#redocly_container'),
@@ -70,7 +81,12 @@ const RedoclyAPIBlock = ({ src }) => {
 };
 
 RedoclyAPIBlock.propTypes = {
-  src: PropTypes.string
+  src: PropTypes.string,
+  width: PropTypes.string,
+  typography: PropTypes.string,
+  codeBlock: PropTypes.string,
+  disableSidebar: PropTypes.bool,
+  disableSearch: PropTypes.bool,
 };
 
 export { RedoclyAPIBlock };